# Conflicts:
#	lib/lf/mesh/utils/structured_mesh_builder.h

+++ lib/lf/geometry/tria_o2.cc
@@ -63,7 +63,7 @@ Eigen::MatrixXd TriaO2::Jacobian(const Eigen::MatrixXd& local) const {
 
   Eigen::MatrixXd tmp(gamma_.rows(), 2 * local.cols());
 
-  for (int i = 0; i < local.cols(); ++i) {
+  for (long i = 0; i < local.cols(); ++i) {
     tmp.block(0, 2 * i, tmp.rows(), 2) =
         gamma_x_2_.array().rowwise() * local.col(i).transpose().array();
   }
@@ -89,14 +89,14 @@ Eigen::MatrixXd TriaO2::JacobianInverseGramian(
   Eigen::MatrixXd jacInvGram(jac.rows(), jac.cols());
 
   if (DimGlobal() == 2) {
-    for (int i = 0; i < local.cols(); ++i) {
+    for (long i = 0; i < local.cols(); ++i) {
       auto jacobian = jac.block(0, 2 * i, 2, 2);
       jacInvGram.block(0, 2 * i, 2, 2) << jacobian(1, 1), -jacobian(1, 0),
           -jacobian(0, 1), jacobian(0, 0);
       jacInvGram.block(0, 2 * i, 2, 2) /= jacobian.determinant();
     }
   } else {
-    for (int i = 0; i < local.cols(); ++i) {
+    for (long i = 0; i < local.cols(); ++i) {
       auto jacobian = jac.block(0, 2 * i, jac.rows(), 2);
 
       auto A = jacobian.col(0);
@@ -124,11 +124,11 @@ Eigen::VectorXd TriaO2::IntegrationElement(const Eigen::MatrixXd& local) const {
   Eigen::VectorXd intElem(local.cols());
 
   if (DimGlobal() == 2) {
-    for (int i = 0; i < local.cols(); ++i) {
+    for (long i = 0; i < local.cols(); ++i) {
       intElem(i) = std::abs(jac.block(0, 2 * i, 2, 2).determinant());
     }
   } else {
-    for (int i = 0; i < local.cols(); ++i) {
+    for (long i = 0; i < local.cols(); ++i) {
       auto jacobian = jac.block(0, 2 * i, jac.rows(), 2);
 
       auto A = jacobian.col(0);
@@ -160,7 +160,9 @@ std::unique_ptr<Geometry> TriaO2::SubGeometry(dim_t codim, dim_t i) const {
       LF_ASSERT_MSG(0 <= i && i <= 5, "i is out of bounds");
       return std::make_unique<Point>(coords_.col(i));
     }
-    default: { LF_VERIFY_MSG(false, "codim is out of bounds") }
+    default: {
+      LF_VERIFY_MSG(false, "codim is out of bounds")
+    }
   }
 }
 
@@ -227,7 +229,9 @@ std::vector<std::unique_ptr<Geometry>> TriaO2::ChildGeometry(
 
         break;
       }
-      default: { LF_VERIFY_MSG(false, "Illegal co-dimension"); }
+      default: {
+        LF_VERIFY_MSG(false, "Illegal co-dimension");
+      }
     }
   }
